+++ .gitignore
@@ -1,2 +1,2 @@
-app/__pycache__
+__pycache__
 app/secret.json
