+++ app/late_investigation/models/model.py
@@ -1,0 +1,5 @@
+from django.db import models
+
+class Model(models.Model):
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
